requirements.txt
flask==3.1.1
python-telegram-bot==21.10
gunicorn==23.0.0
python-dotenv==1.1.0
psycopg2-binary==2.9.10
